requirements.txt
arch>=5.3.0
yfinance>=0.2.0
pandas>=1.5.0
numpy>=1.23.0
google-cloud-bigquery>=3.0.0
Flask>=3.0.0
functions-framework>=3.0.0
requests>=2.31.0
google-generativeai>=0.3.0
reportlab>=4.0.0
chromadb>=0.4.0
google-cloud-storage>=2.0.0

# Binance API Integration
python-binance>=1.0.19
python-dotenv>=1.0.0
